+++ src/py_bugger/cli_messages.py
@@ -1,4 +1,9 @@
 """Messages longer than one line."""
 
-msg_bare_call = """A bare py-bugger call makes no changes to your project.
-You must be explicit about what kinds of errors you want to induce in the project."""
+msg_exception_type_required = (
+    "You must be explicit about what kinds of errors you want to induce in the project."
+)
+
+msg_target_file_dir = (
+    "Target file overrides target dir. Please only pass one of these args."
+)